refactor: Split large test files into focused modules by functionality
