+++ tests/api_client/test_api_methods.py
@@ -1,18 +1,10 @@
-"""Tests for the NoaaApiClient class."""
+"""Tests for NoaaApiClient API methods functionality."""
 
-import time
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 import pytest
-from requests.exceptions import (
-    ConnectionError,
-    HTTPError,
-    JSONDecodeError,
-    RequestException,
-    Timeout,
-)
 
-from accessiweather.api_client import LOCATION_TYPE_COUNTY, NoaaApiClient, NoaaApiError
+from accessiweather.api_client import NoaaApiClient
 
 # Sample test data
 SAMPLE_POINT_DATA = {
@@ -178,96 +170,11 @@ SAMPLE_HOURLY_FORECAST_DATA = {
 }
 
 
-# Fixture to create a NoaaApiClient instance
 @pytest.fixture
 def api_client():
     return NoaaApiClient(
         user_agent="TestClient", contact_info="test@example.com", enable_caching=False
     )
-
-
-# Fixture to create a NoaaApiClient instance with caching enabled
-@pytest.fixture
-def cached_api_client():
-    return NoaaApiClient(
-        user_agent="TestClient", contact_info="test@example.com", enable_caching=True, cache_ttl=300
-    )
-
-
-@pytest.mark.unit
-def test_init_basic():
-    """Test basic initialization without caching."""
-    client = NoaaApiClient(user_agent="TestClient")
-
-    assert client.user_agent == "TestClient"
-    assert client.headers["User-Agent"] == "TestClient (TestClient)"
-    assert client.cache is None
-
-
-@pytest.mark.unit
-def test_init_with_contact():
-    """Test initialization with contact info."""
-    client = NoaaApiClient(user_agent="TestClient", contact_info="test@example.com")
-
-    assert client.headers["User-Agent"] == "TestClient (test@example.com)"
-
-
-@pytest.mark.unit
-def test_init_with_caching():
-    """Test initialization with caching enabled."""
-    client = NoaaApiClient(user_agent="TestClient", enable_caching=True, cache_ttl=300)
-
-    assert client.cache is not None
-    assert client.cache.default_ttl == 300
-
-
-@pytest.mark.unit
-@pytest.mark.api
-def test_get_point_data_success(api_client):
-    """Test getting point data successfully."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        result = api_client.get_point_data(lat, lon)
-
-        assert result == SAMPLE_POINT_DATA
-        mock_get.assert_called_once()
-        assert f"points/{lat},{lon}" in mock_get.call_args[0][0]
-
-
-@pytest.mark.unit
-@pytest.mark.api
-def test_get_point_data_cached(cached_api_client):
-    """Test that point data is cached."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        # First call should hit the API
-        result1 = cached_api_client.get_point_data(lat, lon)
-        # Second call should use cache
-        result2 = cached_api_client.get_point_data(lat, lon)
-
-        assert result1 == result2
-        mock_get.assert_called_once()
-
-
-def test_get_point_data_force_refresh(cached_api_client):
-    """Test that force_refresh bypasses cache."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        # First call
-        cached_api_client.get_point_data(lat, lon)
-        # Second call with force_refresh
-        cached_api_client.get_point_data(lat, lon, force_refresh=True)
-
-        assert mock_get.call_count == 2
 
 
 def test_get_forecast_success(api_client):
@@ -308,19 +215,6 @@ def test_get_forecast_no_url(api_client):
             api_client.get_forecast(lat, lon)
 
         assert "Could not find forecast URL" in str(exc_info.value)
-
-
-def test_identify_location_type_county(api_client):
-    """Test identifying county location type."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        location_type, location_id = api_client.identify_location_type(lat, lon)
-
-        assert location_type == LOCATION_TYPE_COUNTY
-        assert location_id == "PAC091"
 
 
 def test_get_alerts_precise_location(api_client):
@@ -419,187 +313,6 @@ def test_get_national_product_success(api_client):
 
         assert result == SAMPLE_NATIONAL_PRODUCT_TEXT["productText"]
         assert mock_get.call_count == 2
-
-
-def test_rate_limiting(api_client):
-    """Test that requests are rate limited."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        start_time = time.time()
-        api_client.get_point_data(lat, lon)
-        api_client.get_point_data(lat, lon)
-        end_time = time.time()
-
-        # Should have waited at least min_request_interval
-        assert end_time - start_time >= api_client.min_request_interval
-
-
-def test_http_error_handling_404(api_client):
-    """Test handling of 404 Not Found errors."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        # Create a proper mock response with status_code
-        mock_response = MagicMock()
-        mock_response.status_code = 404
-        mock_response.text = "Not Found"
-        mock_response.json.side_effect = JSONDecodeError("Invalid JSON", "", 0)
-        # Create a proper HTTPError with a response attribute
-        http_error = HTTPError("404 Client Error")
-        http_error.response = mock_response
-        mock_get.return_value.raise_for_status.side_effect = http_error
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Resource not found" in str(error)
-        assert error.status_code == 404
-        assert error.error_type == NoaaApiError.CLIENT_ERROR
-
-
-def test_http_error_handling_429(api_client):
-    """Test handling of 429 Rate Limit errors."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        # Create a proper mock response with status_code
-        mock_response = MagicMock()
-        mock_response.status_code = 429
-        mock_response.text = "Too Many Requests"
-        mock_response.json.side_effect = JSONDecodeError("Invalid JSON", "", 0)
-        # Create a proper HTTPError with a response attribute
-        http_error = HTTPError("429 Client Error")
-        http_error.response = mock_response
-        mock_get.return_value.raise_for_status.side_effect = http_error
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Rate limit exceeded" in str(error)
-        assert error.status_code == 429
-        assert error.error_type == NoaaApiError.RATE_LIMIT_ERROR
-
-
-def test_http_error_handling_500(api_client):
-    """Test handling of 500 Server Error."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        # Create a proper mock response with status_code
-        mock_response = MagicMock()
-        mock_response.status_code = 500
-        mock_response.text = "Internal Server Error"
-        mock_response.json.side_effect = JSONDecodeError("Invalid JSON", "", 0)
-        # Create a proper HTTPError with a response attribute
-        http_error = HTTPError("500 Server Error")
-        http_error.response = mock_response
-        mock_get.return_value.raise_for_status.side_effect = http_error
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Server error" in str(error)
-        assert error.status_code == 500
-        assert error.error_type == NoaaApiError.SERVER_ERROR
-
-
-def test_json_decode_error_handling(api_client):
-    """Test handling of JSON decode errors."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.side_effect = JSONDecodeError("Invalid JSON", "", 0)
-        mock_get.return_value.text = "Invalid JSON response"
-        mock_get.return_value.status_code = 200
-        mock_get.return_value.raise_for_status.return_value = None
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Failed to decode JSON response" in str(error)
-        assert error.status_code == 200
-        assert error.error_type == NoaaApiError.PARSE_ERROR
-
-
-def test_request_exception_handling(api_client):
-    """Test handling of request exceptions."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.side_effect = RequestException("Connection error")
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Network error during API request" in str(error)
-        assert error.error_type == NoaaApiError.NETWORK_ERROR
-
-
-def test_timeout_error_handling(api_client):
-    """Test handling of timeout errors."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.side_effect = Timeout("Request timed out")
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Request timed out" in str(error)
-        assert error.error_type == NoaaApiError.TIMEOUT_ERROR
-
-
-def test_connection_error_handling(api_client):
-    """Test handling of connection errors."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get:
-        mock_get.side_effect = ConnectionError("Connection refused")
-
-        with pytest.raises(NoaaApiError) as exc_info:
-            api_client.get_point_data(lat, lon)
-
-        error = exc_info.value
-        assert "Connection error" in str(error)
-        assert error.error_type == NoaaApiError.CONNECTION_ERROR
-
-
-def test_thread_safety(api_client):
-    """Test thread safety of request handling."""
-    import threading
-
-    results = []
-    errors = []
-
-    def make_request():
-        try:
-            result = api_client.get_point_data(40.0, -75.0)
-            results.append(result)
-        except Exception as e:
-            errors.append(e)
-
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        # Create multiple threads making concurrent requests
-        threads = [threading.Thread(target=make_request) for _ in range(5)]
-
-        # Start all threads
-        for thread in threads:
-            thread.start()
-
-        # Wait for all threads to complete
-        for thread in threads:
-            thread.join()
-
-        # Check results
-        assert len(errors) == 0
-        assert all(result == SAMPLE_POINT_DATA for result in results)
-        # Each request should have been made with proper rate limiting
-        assert mock_get.call_count == 5
 
 
 def test_get_stations_success(api_client):
@@ -717,19 +430,3 @@ def test_get_hourly_forecast_no_url(api_client):
             api_client.get_hourly_forecast(lat, lon)
 
         assert "Could not find hourly forecast URL" in str(exc_info.value)
-
-
-def test_rate_limiting_multiple_endpoints(api_client):
-    """Test that requests to multiple endpoints are rate limited."""
-    lat, lon = 40.0, -75.0
-    with patch("requests.get") as mock_get, patch("time.sleep") as mock_sleep:
-        # Set up the mock response
-        mock_get.return_value.json.return_value = SAMPLE_POINT_DATA
-        mock_get.return_value.raise_for_status.return_value = None
-
-        # Make multiple requests to the same endpoint
-        api_client.get_point_data(lat, lon)
-        api_client.get_point_data(lat, lon)
-
-        # Verify that sleep was called at least once for rate limiting
-        assert mock_sleep.call_count > 0
