+++ tests/api_wrapper/__init__.py
@@ -1,0 +1,1 @@
+"""API wrapper tests package."""
