+++ tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Test fixtures package."""
