+++ tests/gui/ui_manager/__init__.py
@@ -1,0 +1,1 @@
+"""UI Manager test package."""
